requirements.txt
fastapi[all]
uvicorn>=0.22.0
pydantic>=2.5.2
python-dotenv>=0.21.1
motor>=3.1.2